+++ lib/l10n/app_localizations_es.dart
@@ -4399,6 +4399,14 @@ class AppLocalizationsEs extends AppLocalizations {
   String get syncSkip => 'Omitir';
 
   @override
+  String get syncStepBackfillAgentEntityClocks =>
+      'Rellenar relojes de entidades de agente';
+
+  @override
+  String get syncStepBackfillAgentLinkClocks =>
+      'Rellenar relojes de enlaces de agente';
+
+  @override
   String get syncStepAgentEntities => 'Entidades de agente';
 
   @override