+++ lib/l10n/app_localizations_fr.dart
@@ -4401,6 +4401,14 @@ class AppLocalizationsFr extends AppLocalizations {
   String get syncSkip => 'Ignorer';
 
   @override
+  String get syncStepBackfillAgentEntityClocks =>
+      'Remplir les horloges des entités d\'agent';
+
+  @override
+  String get syncStepBackfillAgentLinkClocks =>
+      'Remplir les horloges des liens d\'agent';
+
+  @override
   String get syncStepAgentEntities => 'Entités d\'agent';
 
   @override